requirements.txt
awesome-slugify==1.6.5

Django==2.2.3

django-ckeditor==5.7.1

django-js-asset==1.2.2

django-password-reset==2.0

Pillow==6.1.0

pytz==2019.1

redis==3.3.8

regex==2019.8.19

sqlparse==0.3.0

Unidecode==0.4.21

uWSGI==2.0.18

virtualenv==16.6.2
